Refactored LLM-related utils and resolved sponsor logo visibility issue in README.md for users with white themes.
