+++ README.md
@@ -32,7 +32,15 @@ Development of this free, open-source project is made possible by all the <a hre
 </p>
 
 <p align="center">
-<a href="https://www.influxdata.com/"><img src="https://user-images.githubusercontent.com/22963551/232278272-24712932-7bb3-4fbe-b7e9-4b649b8104b2.png" height="40px" alt="InfluxData" /></a>&nbsp;&nbsp;&nbsp;&nbsp;<a href="https://roost.ai"><img src="https://roost.ai/hubfs/logos/Roost.ai-logo-gold.svg" height="40px" alt="Roost.AI" /></a>&nbsp;&nbsp;&nbsp;&nbsp;<a href="https://nuclei.ai/"><img src="https://nuclei.ai/images/nuclei_logo_with_text_white.svg" height="40px" alt="NucleiAI" /></a>&nbsp;&nbsp;&nbsp;&nbsp;<a href="https://www.algohash.org/"><img src="https://user-images.githubusercontent.com/22963551/232277567-78d642c1-ce21-4bcc-a209-c2a23495ab2d.png" height="40px" alt="AlgohashFe" /></a>&nbsp;&nbsp;&nbsp;&nbsp;</p>
+<a href="https://www.influxdata.com/"><img src="https://influxdata.github.io/branding/img/downloads/influxdata-logo--full--castle-alpha.png" height="40px" alt="InfluxData" style="border: 1px solid black;"/></a>
+&nbsp;&nbsp;&nbsp;&nbsp;
+<a href="https://roost.ai"><img src="https://roost.ai/hubfs/logos/Roost.ai-logo-gold.svg" height="40px" alt="Roost.AI" style="border: 1px solid black;"/></a>
+&nbsp;&nbsp;&nbsp;&nbsp;
+<a href="https://nuclei.ai/"><img src="https://theme.zdassets.com/theme_assets/9187008/4c65a25e4adfb45e973202bfeda2da78c8bb892a.png" height="40px" alt="NucleiAI" style="border: 1px solid black;"/></a>
+&nbsp;&nbsp;&nbsp;&nbsp;
+<a href="https://www.algohash.org/"><img src="https://user-images.githubusercontent.com/22963551/232277567-78d642c1-ce21-4bcc-a209-c2a23495ab2d.png" height="40px" alt="AlgohashFe" style="border: 1px solid black;"/></a>&nbsp;&nbsp;&nbsp;&nbsp;
+<a href="https://www.turintech.ai/"><img src="https://miro.medium.com/v2/resize:fit:444/1*esrNX1hPdZ2Ww84fExGdOg.png" height="40px" alt="AlgohashFe" style="border: 1px solid black;"/></a>&nbsp;&nbsp;&nbsp;&nbsp;
+</p>
 
 <p align="center">
 <a href="https://github.com/robinicus"><img src="https://github.com/robinicus.png" width="50px" alt="robinicus" /></a>&nbsp;&nbsp;<a href="https://github.com/prompthero"><img src="https://github.com/prompthero.png" width="50px" alt="prompthero" /></a>&nbsp;&nbsp;<a href="https://github.com/crizzler"><img src="https://github.com/crizzler.png" width="50px" alt="crizzler" /></a>&nbsp;&nbsp;<a href="https://github.com/tob-le-rone"><img src="https://github.com/tob-le-rone.png" width="50px" alt="tob-le-rone" /></a>&nbsp;&nbsp;<a href="https://github.com/FSTatSBS"><img src="https://github.com/FSTatSBS.png" width="50px" alt="FSTatSBS" /></a>&nbsp;&nbsp;<a href="https://github.com/toverly1"><img src="https://github.com/toverly1.png" width="50px" alt="toverly1" /></a>&nbsp;&nbsp;<a href="https://github.com/ddtarazona"><img src="https://github.com/ddtarazona.png" width="50px" alt="ddtarazona" /></a>&nbsp;&nbsp;<a href="https://github.com/Nalhos"><img src="https://github.com/Nalhos.png" width="50px" alt="Nalhos" /></a>&nbsp;&nbsp;<a href="https://github.com/Kazamario"><img src="https://github.com/Kazamario.png" width="50px" alt="Kazamario" /></a>&nbsp;&nbsp;<a href="https://github.com/pingbotan"><img src="https://github.com/pingbotan.png" width="50px" alt="pingbotan" /></a>&nbsp;&nbsp;<a href="https://github.com/indoor47"><img src="https://github.com/indoor47.png" width="50px" alt="indoor47" /></a>&nbsp;&nbsp;<a href="https://github.com/AuroraHolding"><img src="https://github.com/AuroraHolding.png" width="50px" alt="AuroraHolding" /></a>&nbsp;&nbsp;<a href="https://github.com/kreativai"><img src="https://github.com/kreativai.png" width="50px" alt="kreativai" /></a>&nbsp;&nbsp;<a href="https://github.com/hunteraraujo"><img src="https://github.com/hunteraraujo.png" width="50px" alt="hunteraraujo" /></a>&nbsp;&nbsp;<a href="https://github.com/Explorergt92"><img src="https://github.com/Explorergt92.png" width="50px" alt="Explorergt92" /></a>&nbsp;&nbsp;<a href="https://github.com/judegomila"><img src="https://github.com/judegomila.png" width="50px" alt="judegomila" /></a>&nbsp;&nbsp;
