+++ autogpt/chat.py
@@ -2,7 +2,7 @@ import time
 
 from openai.error import RateLimitError
 
-from autogpt import token_counter
+import autogpt.llm_utils
 from autogpt.config import Config
 from autogpt.llm_utils import create_chat_completion
 from autogpt.logs import logger
@@ -40,7 +40,7 @@ def generate_context(prompt, relevant_memory, full_message_history, model):
     next_message_to_add_index = len(full_message_history) - 1
     insertion_index = len(current_context)
     # Count the currently used tokens
-    current_tokens_used = token_counter.count_message_tokens(current_context, model)
+    current_tokens_used = autogpt.llm_utils.count_message_tokens(current_context, model)
     return (
         next_message_to_add_index,
         current_tokens_used,
@@ -106,7 +106,7 @@ def chat_with_ai(
                     prompt, relevant_memory, full_message_history, model
                 )
 
-            current_tokens_used += token_counter.count_message_tokens(
+            current_tokens_used += autogpt.llm_utils.count_message_tokens(
                 [create_chat_message("user", user_input)], model
             )  # Account for user input (appended later)
 
@@ -114,7 +114,7 @@ def chat_with_ai(
                 # print (f"CURRENT TOKENS USED: {current_tokens_used}")
                 message_to_add = full_message_history[next_message_to_add_index]
 
-                tokens_to_add = token_counter.count_message_tokens(
+                tokens_to_add = autogpt.llm_utils.count_message_tokens(
                     [message_to_add], model
                 )
                 if current_tokens_used + tokens_to_add > send_token_limit:
